refactor(callbacks): split register_callbacks — Overview + US Grid (Step 10a/10b)
First step of the ``register_callbacks`` decomposition (issue #87,
Step 10). Begins moving the Dash decorator block from the 2,000-line
``register_callbacks`` function in callbacks.py into per-tab
``register_X_callbacks(app)`` functions that live in the tab modules.

After every step lands, each tab module owns its read path end-to-end
— layout (``tab_X.py``), helpers (already extracted in #98–#108), and
callback wiring all in coherent places.

## Why incremental

``register_callbacks`` is 32 callbacks across 18 section headers and
mixes per-tab and cross-cutting concerns (data loading, persona
switching, URL state, NEXD-* features). Some sections aren't cleanly
tab-bounded. A big-bang split has real risk: each callback's
``Input``/``Output`` ids must keep resolving against the layout, and
the Dash callback registry checks for duplicate outputs across the
whole app.

Splitting incrementally — cleanest tabs first, tricky cross-cutting
callbacks last (or never) — lets each PR stay reviewable and
revert-able.

## What moves in this PR

Two tabs with the cleanest callback boundaries:

* **Overview** — 1 callback (``update_overview_tab``, 65 lines) moves
  to ``register_overview_callbacks(app)`` in
  ``components/_callbacks_overview.py``.
* **US Grid** — 3 callbacks (``update_us_grid_snapshot``,
  ``drilldown_from_us_grid``, ``drilldown_from_us_grid_map``, 127
  lines total) move to ``register_us_grid_callbacks(app)`` in
  ``components/_callbacks_us_grid.py``.

``register_callbacks`` now calls these two functions in place of the
inline decorator blocks. The section comments stay in callbacks.py as
breadcrumbs.

## Public-import surface preserved

callbacks.py re-imports both ``register_X_callbacks`` and the helpers
they call. Helpers picked up new ``# noqa: F401 — re-export`` markers
since the call sites moved out of callbacks.py itself.

The two new ``register_X_callbacks(app)`` functions follow the same
shape:
* Define lazy imports inside the function body for symbols only the
  callbacks need (e.g. ``build_metrics_bar``, ``REGION_GROUPS``).
* Define ``@app.callback`` decorators directly, just like
  ``register_callbacks`` did before.
* Closure semantics unchanged — the callbacks only capture ``app``,
  no other locals from the outer function.

## Test patches

No patches updated. The callback registration tests don't introspect
the decorator block's location; they just check that
``register_callbacks(app)`` produces the expected ``callback_map``.

## Size impact

callbacks.py: 2,255 → 2,076 (–179 lines)
_callbacks_overview.py: 2,057 → 2,135 (+78)
_callbacks_us_grid.py: 720 → 842 (+122)

## What's next

* **Step 10c** — Forecast callbacks (TAB 1: Demand Forecast at
  line 708, FORECAST TAB R4a at line 1643, FORECAST REPLAY at line
  2103). Largest remaining tab — ~750 lines across three sections.
* **Step 10d** — Generation TAB 4 + Scenario TAB 6 (likely hidden
  per the v2 shell; may be candidates for deletion vs migration).
* **Final** — update ``docs/internal/CALLBACKS_DECOMPOSITION_PLAN.md``
  with the final state.

Cross-cutting callbacks (data loading, persona, URL state, NEXD-*
features) stay in callbacks.py — they're foundational rather than
tab-specific.

## Tests

All 1,573 unit + integration tests pass locally. Ruff lint + format
clean.
